Add Observations structure for handling data
Contain both the data and the distribution to be fitted, can be used to carry around all the information needed in a
neat way. Will be build based on user data (with automatic detection for some case, e.g. simple unweighted graphs
will be assigned a Bernoulli) and then pass around internally.